Fix the memory leak in case of slow consumer
Background: If we have routes flapping quickly (e,g, link flaps),
due to the slow consumer at orchagent, the redis-DB would be consuming
more and more memory, and eventually could eat the whole box memory and
cause OOM kill of database.

We have partial fix in: https://rb.corp.linkedin.com/r/1550031/

However, it does not completely address the problem. The problem was that
even we have one notification per key (compressed vs per request),
since we are using pops at consumer side, one notification would mostly
pop all keys at one shot and the notifications is significant higher than
the pops operations, and the channel could still build up and cause memory leak.

The change here is to publish one notification if the keyset was empty before
at producer side, and consumer size we always pops all keys. In LinkedIn deployment
the size of keyset is limited (< 8K), and even if keyset is more than that, it just
take orchagent long to process a batch.
The tests showed no memory build up for long time (several hours) with the same
condition as before.

Changes for pop operation is done to make UT tests passing. However, current
SONiC is not using pop, so it is not applicable for production.

More test result details see:
https://jira01.corp.linkedin.com:8443/browse/FALCO-1576

Signed-off-by: Zhenggen Xu <zxu@linkedin.com>

Addtional caution for notification and small optimization
in consumer side to not SPOP at all when keyset is empty.

Signed-off-by: Zhenggen Xu <zxu@linkedin.com>

    RB=1551480
    G=lnos-reviewers
    R=pchaudhary,pmao,rmolina,samaity,sfardeen,zxu
    A=

     Conflicts:
            common/consumer_state_table_pops.lua
	    common/consumerstatetable.h
	    common/consumerstatetable.h
	    common/consumertablebase.h
	    common/producerstatetable.cpp